currency converter
web service is a program that, when you send it web requests,
automatically generates a web page with the information that you asked
for. In our case, the web service will tell us the current exchange
rate for most of the major international currencies. Use
string-manipulation methods to read the web page and extract the exact
information we need.